✨ add : NewProductDetail user set category
